src/main/java/br/com/fogliato/arrays/GroupAnagrams.java
package br.com.fogliato.arrays;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

// https://leetcode.com/problems/group-anagrams/description/
public class GroupAnagrams {

    public List<List<String>> groupAnagrams(String[] strs) {
        if (strs.length == 0) {
            return new ArrayList();
        }

        Map<String, List<String>> anagrams = new HashMap<>();

        for (String s : strs) {
            int[] count = new int[26];
            for (char c : s.toCharArray()) {
                count[c - 'a']++;
            }

            StringBuilder sb = new StringBuilder("");
            for (int i = 0; i < 26; i++) {
                sb.append(count[i]);
            }
            String key = sb.toString();
            if (!anagrams.containsKey(key)) {
                anagrams.put(key, new ArrayList());
            }
            anagrams.get(key).add(s);
        }
        return new ArrayList(anagrams.values());
    }
}
